Alternating Direction Method of Multipliers (ADMM)

PowerModelsADA.solve_dopf_admm — Function

solve_dopf_admm(data::Dict{String, <:Any}, model_type::DataType, optimizer; 
mismatch_method::String="norm", tol::Float64=1e-4, max_iteration::Int64=1000, 
print_level::Int64=1, print_optimizer_info::Bool=false, alpha::Real=1000)

Solve the distributed OPF problem using ADMM algorithm.

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

model_type::DataType : power flow formulation (PowerModel type)

optimizer : optimizer JuMP initiation object

mismatch_method::String="norm" : mismatch calculation method (norm, max)

tol::Float64=1e-4 : mismatch tolerance

tol_dual::Float64=1e-4 : dual residual tolerance

max_iteration::Int64=1000 : maximum number of iteration

print_level::Int64=1 : 0 - no print, 1 - print mismatch after each iteration and result summary, 2 - print optimizer output

alpha::Real=1000 : algorithm parameter

PowerModelsADA.solve_dopf_admm_coordinated — Function

solve_dopf_admm_coordinated(data::Dict{String, <:Any}, model_type::DataType, optimizer; tol::Float64=1e-4, 
max_iteration::Int64=1000, print_level::Int64=1, alpha::Real=1000)

Solve the distributed OPF problem using ADMM algorithm with central coordinator.

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

model_type::DataType : power flow formulation (PowerModel type)

optimizer : optimizer JuMP initiation object

mismatch_method::String="norm" : mismatch calculation method (norm, max)

tol::Float64=1e-4 : mismatch tolerance

max_iteration::Int64=1000 : maximum number of iteration

print_level::Int64=1 : 0 - no print, 1 - print mismatch after each iteration and result summary, 2 - print optimizer output

alpha::Real=1000 : algorithm parameters

PowerModelsADA.admm_methods — Module

ADMM algorithm module contains build and update methods

PowerModelsADA.admm_methods.build_method — Method

build PowerModel object for the ADMM algorithm

PowerModelsADA.admm_methods.initialize_method — Method

initialize the ADMM algorithm

PowerModelsADA.admm_methods.objective_admm — Method

ADMM algorithm objective function

PowerModelsADA.admm_methods.solve_method — Method

solve distributed OPF using ADMM algorithm

PowerModelsADA.admm_methods.update_method — Method

update the ADMM algorithm data after each iteration

PowerModelsADA.admm_coordinated_methods — Module

ADMM algorithm module containsbuild and update methods

PowerModelsADA.admm_coordinated_methods.build_method_coordinator — Method

build PowerModel object for the ADMM algorithm coordinator

PowerModelsADA.admm_coordinated_methods.build_method_local — Method

build PowerModel object for the ADMM algorithm local area

PowerModelsADA.admm_coordinated_methods.initialize_method_coordinator — Method

initializethe ADMM algorithm coordinator

PowerModelsADA.admm_coordinated_methods.initialize_method_local — Method

initialize the ADMM algorithm local area

PowerModelsADA.admm_coordinated_methods.objective_admm_coordinator — Method

ADMM algorithm objective function of the local area

PowerModelsADA.admm_coordinated_methods.objective_admm_local — Method

ADMM algorithm objective function of the coordinator

PowerModelsADA.admm_coordinated_methods.solve_method — Method

solve distributed OPF using ADMM algorithm with central coordinator

PowerModelsADA.admm_coordinated_methods.update_method_coordinator — Method

update the ADMM algorithm coordinator data after each iteration

PowerModelsADA.admm_coordinated_methods.update_method_local — Method

update the ADMM algorithm coordinator data after each iteration