Data Structure

Input Data

Case

The input data is the same as PowerModels with one essential exeption. The buses in the PowerModels dictionary must contain area key with more than one area number. To load a data file, we use the same PowerModels methods. The PowerModels data can be loaded as follow:

case_path = "test/data/case14.m"
data = parse_file(case_path)

Partitioning

To check the areas in a PowerModels data, use PMADA.get_areas_id(data) to get all areas ids in data. If the data dictionary dosen't contain more than one area, there are two methods to parition the system assign_area! or partition_system!

PMADA.assign_area! — Function

assign area to the PowerModel data using a dictionary with (bus => area) Int pairs

assign area to the PowerModel data using a CVS file with buses and area id

assign area to the PowerModel data using a vector with (bus => area) pairs

assign area to the PowerModel data using a matrix with [bus, area] colmuns or rows

PMADA.partition_system! — Function

partition_system!(data::Dict, n::Int64; configuration::Symbol=:edge_cut, print_info::Bool=false)

Partition a system into n areas using KaHyPar partition algorithm

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

n::Int : number of areas

configuration::Symbol : partition meteric (:edgecut or :connectivity). The defult is :edgecut

print_info::Bool : print partition algorithm information

An example of partition file is shown in parition example.

Before running the distirbuted algorithm, PMADA.jl internally decmpose the original system into subsystems. It decuple the tie-lines by introducing dummy buses and virtual generators at the tie-lines ends. This process is pefromed using decompose_system function.

PMADA.decompose_system — Function

decompose_system(data::Dict{String, <:Any})

decompose a system into areas defined by bus area.

decompose an area with area id

Output Data

The output of the distributed algorithms is stored in a dictionary the keys are the area id and the value are the area data dictionary that contain the results.

Historical Data

TODO

Generation Cost

To calculate the objective function of the central algorithm use calc_dist_gen_cost.

PMADA.calc_dist_gen_cost — Function

calculate distributed solution operation cost

To compar the distributed algorithm objective function value with the central OPF, use compare_solution to get the absolute relative error.

PMADA.compare_solution — Function

compare the distributed algorithm solution with PowerModels centralized solution