Library

PMADA.calc_global_mismatch — Method

calculate the global mismatch based on local mismatch

PMADA.calc_mismatch! — Method

calc_mismatch!(data::Dict{String, <:Any},method::String="norm"; p::Int64=2)

calculate the mismatch using p-norm and return the area data dictionary with the mismatch as seen by the area.

PMADA.initialize_dopf_parameters! — Method

initialize dopf parameters

PMADA.save_solution! — Method

save last solution in previous_solutions vector

PMADA.solve_dopf_coordinated — Method

solve_dopf_coordinated(data::Dict{String, <:Any}, model_type::DataType, optimizer, model_type::DataType; mismatch_method::String="norm", tol::Float64=1e-4, max_iteration::Int64=1000, verbose::Int64=1, kwargs...)

Solve OPF problem using distributed algorithm with central coordinator. The distributed algorithm is defined by the dopf_method Module.

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

model_type::DataType : power flow formulation (PowerModel type)

optimizer : optimizer JuMP initiation object

model_type::DataType : module contains the distributed algorithm methods as follow:

initializemethodcoordinator::Function : initliize distributed algorithm parameters and shared variables

initializemethodlocal::Function : initliize distributed algorithm parameters and shared variables

updatemethodcoordinator::Function : update the algorithm after each iteration

updatemethodlocal::Function : update the algorithm after each iteration

buildmethodcoordinator::Function : problem formulation

buildmethodlocal::Function : problem formulation

mismatch_method::String="norm" : mismatch calculation method (norm, max)

tol::Float64=1e-4,

max_iteration::Int64=1000,

verbose::Int64=1 : print mismatch after each iteration and result summary

kwargs = distributed algorithm parameters

PMADA.solve_local! — Method

solve local problem and update the solution and shared variables dictionaries

PMADA.update_flag_convergance! — Method

check the shared variables of a local area are within tol

PMADA.update_global_flag_convergance — Method

check the flag convergence for all areas and return a global variables

PMADA.update_iteration! — Method

update iteration

PMADA.update_shared_variable! — Method

update primal variables after obtaining a solution at each iteraton

PMADA.update_solution! — Method

update the area data and local shared variables after obtaining a solution at each iteraton

PMADA.add_virtual_gen! — Method

add virtual geneartors at the neighboring buses of an area

PMADA.arrange_areas_id! — Method

arrange area id from 1 to number of areas this step is necessary when having area number 0 and using central coordinator

PMADA.assign_area! — Method

assign area to the PowerModel data using a dictionary with (bus => area) Int pairs

PMADA.assign_area! — Method

assign area to the PowerModel data using a matrix with [bus, area] colmuns or rows

PMADA.assign_area! — Method

assign area to the PowerModel data using a CVS file with buses and area id

PMADA.assign_area! — Method

assign area to the PowerModel data using a vector with (bus => area) pairs

PMADA.decompose_coordinator — Method

define system coordinator

PMADA.decompose_system — Method

decompose_system(data::Dict{String, <:Any})

decompose a system into areas defined by bus area.

PMADA.decompose_system — Method

decompose an area with area id

PMADA.get_area_id — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_area_id — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_areas_bus — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_areas_bus — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_areas_id — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_areas_id — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_local_bus — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_local_bus — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_neighbor_bus — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_neighbor_bus — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_neighbor_bus — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_neighbor_bus — Method

helper functions to handle area ids, local buses, neighbor buses

PMADA.get_shared_component — Method

get the shared buses and branches between defined area and all other areas

PMADA.get_shared_component — Method

get the shared buses and branches between defined area and all other areas

PMADA.get_shared_component — Method

get the shared buses and branches between defined area and all other areas

PMADA.get_shared_component — Method

get the shared buses and branches between defined area and all other areas

PMADA.receive_shared_data! — Method

store received data in the local data dictionary

PMADA.receive_shared_data! — Method

deserialize and store the received data in the local data

PMADA.send_shared_data — Method

get the shared data with or without serialization

PMADA.constraint_opf — Method

define OPF problem constraints

PMADA.no_objective — Method

no objective function case

PMADA.objective_min_fuel_and_consensus! — Function

define objective function from PowerModels and algorithm specific objective

PMADA.variable_opf — Method

define OPF problem variable

PMADA.calc_dist_gen_cost — Method

calculate distributed solution operation cost

PMADA.compare_solution — Method

compare the distributed algorithm solution with PowerModels centralized solution

PMADA.partition_system! — Method

partition_system!(data::Dict, n::Int64; configuration::Symbol=:edge_cut, print_info::Bool=false)

Partition a system into n areas using KaHyPar partition algorithm

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

n::Int : number of areas

configuration::Symbol : partition meteric (:edgecut or :connectivity). The defult is :edgecut

print_info::Bool : print partition algorithm information

PMADA._initialize_shared_variable — Function

initialize shared variable dictionary

PMADA._var — Method

return JuMP variable object from PowerModel object

PMADA.initial_value — Function

initial_value(data::Dict{String, <:Any}, var::String, idx::Int, method::String="flat")

assign initial value based on initialization method

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

var::String : variable names

idx::Int : variable index

method::String : initialization method ("flat", "previous_solution")

PMADA.initialize_all_variable — Method

initialize_all_variable(data::Dict{String, <:Any}, model_type::DataType)

return a dictionary contains all the problem variables. Can be used to store the last solution

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

model_type::DataType : power flow formulation (PowerModel type)

PMADA.initialize_dual_variable! — Method

initialize the dual shared variables

PMADA.initialize_dual_variable_coordinator! — Method

initialize the dual shared variables for coordinator

PMADA.initialize_dual_variable_local! — Method

initialize the dual shared variables

PMADA.initialize_shared_variable! — Method

initialize the primal shared variables

PMADA.initialize_shared_variable_coordinator! — Method

initialize the primal shared variables for coordinator

PMADA.initialize_shared_variable_local! — Method

initialize the primal shared variables

PMADA.variable_names — Method

idinifiy all the variables names

PMADA.variable_shared_names — Method

idinifiy the shared bus and branch variables names