Library

PowerModelsADA.calc_dual_residual! — Method

calc_dual_residual!(data::Dict{String, <:Any}; central::Bool=false)

calculate the dual redidual as seen by the area. Set central=true if the algorithm uses the optimality condition of a central coordinator.

PowerModelsADA.calc_global_mismatch — Method

calculate the global mismatch based on local mismatch

PowerModelsADA.calc_mismatch! — Method

calc_mismatch!(data::Dict{String, <:Any}; central::Bool=false)

calculate the mismatch and return the area data dictionary with the mismatch as seen by the area. Set central=true if the algorithm uses the optimality condition of a central coordinator.

PowerModelsADA.flag_convergance — Method

check flag convergance using mismatch and dual residual

PowerModelsADA.initialize_dopf! — Method

initialize dopf parameters

PowerModelsADA.print_convergence — Method

print final solution status

PowerModelsADA.print_iteration — Function

print iteration information

PowerModelsADA.save_solution! — Method

save last solution in previous_solutions vector

PowerModelsADA.solve_dopf — Method

solve_dopf(data::Dict{String, <:Any}, model_type::DataType, optimizer, dopf_method::Module; print_level::Int64=1, multiprocessors::Bool=false, mismatch_method::String="norm", tol::Float64=1e-4, max_iteration::Int64=1000, save_data::Vector{String}=[], kwargs...)

Solve OPF problem using fully distributed algorithm.

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

model_type::DataType : power flow formulation (PowerModel type)

optimizer : optimizer JuMP initiation object

dopf_method::Module : module contains the distributed algorithm methods as follows:

initialize_method::Function : initialize the algorithm parameters and shared variables

update_method::Function : update the algorithm after each iteration

build_method::Function : problem formulation

print_level::Int64=1 : 0 - no print, 1 - print mismatch after each iteration and result summary, 2 - print optimizer output

multiprocessors::Bool=false : enable multiprocessors using available workers. Multiprocessors feature requires loading the PowerModelsADA and the optimizer packages on all the processors using @everywhere using <package_name>.

mismatch_method::String="norm" : mismatch calculation method (norm, max)

tol::Float64=1e-4 : mismatch tolerance

max_iteration::Int64=1000 : maximum number of iteration

savedata::Vector{String}=[] : vector contains the keys of the dictionaries to be saved at each iteration in "previoussolution". For example, savedata=["solution", "sharedvariable", "mismatch"]

kwargs = includes algorithm-specific and initialization parameters

PowerModelsADA.solve_dopf_coordinated — Method

solve_dopf_coordinated(data::Dict{String, <:Any}, model_type::DataType, optimizer, dopf_method::Module; print_level::Int64=1, multiprocessors::Bool=false, mismatch_method::String="norm", tol::Float64=1e-4, max_iteration::Int64=1000, save_data::Vector{String}=[], kwargs...)

Solve OPF problem using distributed algorithm with central coordinator.

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

model_type::DataType : power flow formulation (PowerModel type)

optimizer : optimizer JuMP initiation object

dopf_method::Module : module contains the distributed algorithm methods as follows:

initialize_method_local::Function : initialize the local algorithm parameters and shared variables

initialize_method_coordinator::Function : initialize the coordinator algorithm parameters and shared variables

update_method_local::Function : update the local data after each iteration

update_method_coordinator::Function : update the coordinator data after each iteration

build_method_local::Function : local problem formulation

build_method_coordinator::Function : coordinator problem formulation

print_level::Int64=1 : 0 - no print, 1 - print mismatch after each iteration and result summary, 2 - print optimizer output

multiprocessors::Bool=false : enable multiprocessors using available workers. Multiprocessors feature requires loading the PowerModelsADA and the optimizer packages on all the processors using @everywhere using <package_name>.

mismatch_method::String="norm" : mismatch calculation method (norm, max)

tol::Float64=1e-4 : mismatch tolerance

max_iteration::Int64=1000 : maximum number of iteration

savedata::Vector{String}=[] : vector contains the keys of the dictionaries to be saved at each iteration in "previous\solution". For example, savedata=["solution", "sharedvariable", "mismatch"]

kwargs = includes algorithm-specific and initialization parameters

PowerModelsADA.solve_dopf_coordinated_mp — Method

solve_dopf_coordinated_mp(data::Dict{String, <:Any}, model_type::DataType, optimizer, dopf_method::Module; print_level::Int64=1, multiprocessors::Bool=false, mismatch_method::String="norm", tol::Float64=1e-4, max_iteration::Int64=1000, save_data::Vector{String}=[], kwargs...)

Solve OPF problem using distributed algorithm with central coordinator on multiprocessors. Multiprocessors feature requires loading the PowerModelsADA and the optimizer packages on all the processors using @everywhere using <package_name>.

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

model_type::DataType : power flow formulation (PowerModel type)

optimizer : optimizer JuMP initiation object

dopf_method::Module : module contains the distributed algorithm methods as follows:

initialize_method_local::Function : initialize the local algorithm parameters and shared variables

initialize_method_coordinator::Function : initialize the coordinator algorithm parameters and shared variables

update_method_local::Function : update the local data after each iteration

update_method_coordinator::Function : update the coordinator data after each iteration

build_method_local::Function : local problem formulation

build_method_coordinator::Function : coordinator problem formulation

print_level::Int64=1 : 0 - no print, 1 - print mismatch after each iteration and result summary, 2 - print optimizer output

mismatch_method::String="norm" : mismatch calculation method (norm, max)

tol::Float64=1e-4 : mismatch tolerance

max_iteration::Int64=1000 : maximum number of iteration

savedata::Vector{String}=[] : vector contains the keys of the dictionaries to be saved at each iteration in "previous\solution". For example, savedata=["solution", "sharedvariable", "mismatch"]

kwargs = includes algorithm-specific and initialization parameters

PowerModelsADA.solve_dopf_coordinated_sp — Method

solve_dopf_coordinated_sp(data::Dict{String, <:Any}, model_type::DataType, optimizer, dopf_method::Module; print_level::Int64=1, multiprocessors::Bool=false, mismatch_method::String="norm", tol::Float64=1e-4, max_iteration::Int64=1000, save_data::Vector{String}=[], kwargs...)

Solve OPF problem using distributed algorithm with central coordinator on single-processors.

Arguments:

data::Dict{String, <:Any} : dictionary contains case in PowerModel format

model_type::DataType : power flow formulation (PowerModel type)

optimizer : optimizer JuMP initiation object

dopf_method::Module : module contains the distributed algorithm methods as follows:

initialize_method_local::Function : initialize the local algorithm parameters and shared variables

initialize_method_coordinator::Function : initialize the coordinator algorithm parameters and shared variables

update_method_local::Function : update the local data after each iteration

update_method_coordinator::Function : update the coordinator data after each iteration

build_method_local::Function : local problem formulation

build_method_coordinator::Function : coordinator problem formulation

print_level::Int64=1 : 0 - no print, 1 - print mismatch after each iteration and result summary, 2 - print optimizer output

mismatch_method::String="norm" : mismatch calculation method (norm, max)

tol::Float64=1e-4 : mismatch tolerance

max_iteration::Int64=1000 : maximum number of iteration

savedata::Vector{String}=[] : vector contains the keys of the dictionaries to be saved at each iteration in "previous\solution". For example, savedata=["solution", "sharedvariable", "mismatch"]

kwargs = includes algorithm-specific and initialization parameters

PowerModelsADA.solve_dopf_mp — Method

solve_dopf_mp(data::Dict{String, <:Any}, model_type::DataType, optimizer, dopf_method::Module; print_level::Int64=1, mismatch_method::String="norm", tol::Float64=1e-4, max_iteration::Int64=1000, save_data::Vector{String}=[], kwargs...)

Solve OPF problem using fully distributed algorithm on multiprocessors. Multiprocessors feature requires loading the PowerModelsADA and the optimizer packages on all the processors using @everywhere using <package_name>.

Arguments:

data::Dict{Int64, <:Any} : dictionary contains area data in PowerModel format

model_type::DataType : power flow formulation (PowerModel type)

optimizer : optimizer JuMP initiation object

dopf_method::Module : module contains the distributed algorithm methods as follows:

initialize_method::Function : initialize the algorithm parameters and shared variables

update_method::Function : update the algorithm after each iteration

build_method::Function : problem formulation

print_level::Int64=1 : 0 - no print, 1 - print mismatch after each iteration and result summary, 2 - print optimizer output

mismatch_method::String="norm" : mismatch calculation method (norm, max)

tol::Float64=1e-4 : mismatch tolerance

max_iteration::Int64=1000 : maximum number of iteration

savedata::Vector{String}=[] : vector contains the keys of the dictionaries to be saved at each iteration in "previoussolution". For example, savedata=["solution", "sharedvariable", "mismatch"]

kwargs = includes algorithm-specific and initialization parameters

PowerModelsADA.solve_dopf_sp — Method

solve_dopf_sp(data::Dict{String, <:Any}, model_type::DataType, optimizer, dopf_method::Module; print_level::Int64=1, mismatch_method::String="norm", tol::Float64=1e-4, max_iteration::Int64=1000, save_data::Vector{String}=[], kwargs...)

Solve OPF problem using fully distributed algorithm on single-processor.

Arguments:

data::Dict{Int64, <:Any} : dictionary contains area data in PowerModel format

model_type::DataType : power flow formulation (PowerModel type)

optimizer : optimizer JuMP initiation object

dopf_method::Module : module contains the distributed algorithm methods as follows:

initialize_method::Function : initialize the algorithm parameters and shared variables

update_method::Function : update the algorithm after each iteration

build_method::Function : problem formulation

print_level::Int64=1 : 0 - no print, 1 - print mismatch after each iteration and result summary, 2 - print optimizer output

mismatch_method::String="norm" : mismatch calculation method (norm, max)

tol::Float64=1e-4 : mismatch tolerance

max_iteration::Int64=1000 : maximum number of iteration

savedata::Vector{String}=[] : vector contains the keys of the dictionaries to be saved at each iteration in "previoussolution". For example, savedata=["solution", "sharedvariable", "mismatch"]

kwargs = includes algorithm-specific and initialization parameters

PowerModelsADA.update_flag_convergence! — Method

check the shared variables of a local area are within tol

PowerModelsADA.update_global_flag_convergence — Method

check the flag convergence for all areas and return a global variables

PowerModelsADA.update_global_flag_convergence — Method

check the flag convergence for all areas and return a global variables

PowerModelsADA.update_iteration! — Method

update iteration

PowerModelsADA.update_shared_variable! — Method

update primal variables after obtaining a solution at each iteraton

PowerModelsADA.update_solution! — Method

update the area data solution dictionary

PowerModelsADA.add_virtual_gen — Method

add virtual generators at the neighboring buses of an area

PowerModelsADA.arrange_areas_id! — Method

arrange area ID from 1 to number of areas. This step is necessary when having area number 0 and using central coordinator

PowerModelsADA.assign_area! — Method

assign area to the system data using a matrix with [bus, area] columnsor rows

PowerModelsADA.assign_area! — Method

assign area to the system data using a CVS file with buses and area ID

PowerModelsADA.assign_area! — Method

assign area to the system data using a vector with (bus => area) pairs

PowerModelsADA.assign_area! — Method

assign area to the system data using a dictionary with (bus => area) Int pairs

PowerModelsADA.decompose_coordinator — Method

obtain system coordinator data

PowerModelsADA.decompose_system — Method

decompose_system(data::Dict{String, <:Any})

decompose a system into areas defined by bus area.

PowerModelsADA.decompose_system — Method

obtain an area decomposition with area ID

PowerModelsADA.get_area_id — Method

helper function to get the area ID

PowerModelsADA.get_area_id — Method

helper function to get the area ID

PowerModelsADA.get_areas_bus — Method

helper functions to all areas buses in a dicrionary

PowerModelsADA.get_areas_bus — Method

helper functions to all areas buses in a dicrionary

PowerModelsADA.get_areas_id — Method

helper function to get all areas IDs

PowerModelsADA.get_areas_id — Method

helper function to get all areas IDs

PowerModelsADA.get_areas_id — Method

helper function to get all areas IDs

PowerModelsADA.get_local_bus — Method

helper functions to get the area's local buses

PowerModelsADA.get_local_bus — Method

helper functions to get the area's local buses

PowerModelsADA.get_neighbor_bus — Method

helper functions to get the area's neighbor buses

PowerModelsADA.get_neighbor_bus — Method

helper functions to get the area's neighbor buses

PowerModelsADA.get_neighbor_bus — Method

helper functions to get the area's neighbor buses

PowerModelsADA.get_neighbor_bus — Method

helper functions to get the area's neighbor buses

PowerModelsADA.get_shared_component — Method

get the shared buses and branches between defined area and all other areas

PowerModelsADA.get_shared_component — Method

get the shared buses and branches between defined area and all other areas

PowerModelsADA.get_shared_component — Method

get the shared buses and branches between defined area and all other areas

PowerModelsADA.get_shared_component — Method

get the shared buses and branches between an area and all other areas

PowerModelsADA.prepare_shared_data — Method

prepare the shared data with or without serialization

PowerModelsADA.receive_shared_data! — Method

store received data in the local data dictionary

PowerModelsADA.receive_shared_data! — Method

deserialize and store the received data in the local data

PowerModelsADA.constraint_opf — Method

define OPF problem constraints

PowerModelsADA.no_objective — Method

no objective function case

PowerModelsADA.objective_min_fuel_and_consensus! — Function

define objective function using PowerModels and algorithm-specific objective

PowerModelsADA.variable_opf — Method

define OPF problem variable

PowerModelsADA.calc_dist_gen_cost — Method

calculate distributed solution operation cost

PowerModelsADA.calc_number_all_variables — Method

get the number of variables in an area

PowerModelsADA.calc_number_areas_variables — Method

get the number of variables in each area

PowerModelsADA.calc_number_shared_variables — Method

get the number of shared variable in a area

PowerModelsADA.calc_number_system_shared_variables — Method

get the number of shared variable in all area

PowerModelsADA.calc_number_variables — Method

get the number of variables

PowerModelsADA.compare_solution — Method

compare the distributed algorithm solution with PowerModels centralized solution

PowerModelsADA.get_diameter — Method

get the communication network diameter

PowerModelsADA._var — Method

return JuMP variable object from PowerModel object

PowerModelsADA.initial_value — Function

initial_value(variable::String, initialization_method::String="flat")

assign initial value based on initialization method

Arguments:

variable::String : variable names

initializationmethod::String="flat : ("flat", "previoussolution")

PowerModelsADA.initialize_all_variable — Function

initialize_all_variable(data::Dict{String, <:Any}, model_type::DataType, dics_name::String="solution", initialization_method::String="flat")

return a dictionary contains all the problem variables. can be used to store the solutions.

Arguments:

data::Dict{String, <:Any} : area data

model_type::DataType : power flow formulation (PowerModel type)

dics_name::String="solution" : location of existing dicrionary to be used to worm start the output

initialization_method::String="flat" : "flat" or "worm" initialization

PowerModelsADA.initialize_shared_variable — Function

initialize shared variable dictionary

PowerModelsADA.variable_names — Method

identifyall the variables names

PowerModelsADA.variable_shared_names — Method

identifythe shared bus and branch variables names